Struct jsonrpc_v1::Response

pub struct Response {
    pub result: Option<Json>,
    pub error: Option<Json>,
    pub id: Json,
}

A JSONRPC response object

Fields

result

The Object that was returned by the invoked method. This must be null in case there was a error invoking the method

error

An Error object if there was an error invoking the method. It must be null if there was no error

id

This must be the same id as the request it is responding to

Methods

impl Response

fn result<T: Deserialize>(&self) -> Result<T, Error>

Extract the result from a response

fn into_result<T: Deserialize>(self) -> Result<T, Error>

Extract the result from a response, consuming the response

fn check_error(self) -> Result<(), Error>

Return the RPC error, if there was one, but do not check the result

fn is_none(&self) -> bool

Returns whether or not the result field is empty